==> notebooks/01_Scraping_Rapper's_names.ipynb <==
# Scraping Ranker.com to get a list of rappers name

In order to scrape genius.com for rap lyrics. I need to have a list of rappers names that I can feed into the api. I can create one manually, but it's easier just to scrape a list of names off the internet. Ranker.com is a public ranking site, and I am choosing to scrape my list from the publically ranked list of the greatest rappers of all time.

### Importing requests, beatiful soup, OS, and selenium for web scraping

from urllib.request import Request, urlopen, quote # Python 3
from bs4 import BeautifulSoup
import requests
import json
import os
import pickle

import time
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.common.exceptions import NoSuchElementException
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.support.wait import WebDriverWait
from selenium.webdriver.common.by import By

import lyricsgenius as genius
client_access_token = "OWyOJwsI5mCr4YxTeADx2ioI1zRV-A8CKUPBxo3GRIyixhVHrDgPfAknjAgwRkyy"
api = genius.Genius(client_access_token)

### Creating a function that will click the load more button until it's gone and scrape the resulting list

In this function, I use selenium to click the load more button until it dissapears, I then store the list of rappers names in a list ans 

def getListFromRanker_Selenium(url):
    browser = webdriver.Chrome()

    browser.get("https://www.ranker.com/crowdranked-list/the-greatest-rappers-of-all-time")
    time.sleep(1)

    elem = browser.find_element_by_tag_name("body")
    element = browser.find_element_by_xpath("//body/button[contains(@id, 'pagination')]")
    i=500
    while i:   
        try:   
            browser.execute_script("arguments[0].click();", element)
            elem.send_keys(Keys.PAGE_DOWN)
            time.sleep(0.4)
            i-=1
        except Exception as e:
            print(e)
            break
            
    print("Complete")
    time.sleep(10)

    post_elems = browser.find_elements_by_class_name("listItem__data")
    artist_names = [p.find_element_by_class_name("listItem__title").text for p in post_elems[:-1]]
    browser.quit()
    return artist_names

url = "https://www.ranker.com/crowdranked-list/the-greatest-rappers-of-all-time"

artist_names = getListFromRanker_Selenium(url)

len(artist_names)

with open('./artist_names.pkl', 'wb+') as f:
    pickle.dump(artist_names, f)
