==> notebooks/02_Scraping_lyrics.ipynb <==
# Scraping lyrics from genius.com

In this notebook, I will create a for loop that makes use of the genius api wrapper and collects all the data from genius for each rapper name that I provide.

import pickle
import json
import time
import pandas as pd
import numpy as np
import lyricsgenius as genius
client_access_token = "OWyOJwsI5mCr4YxTeADx2ioI1zRV-A8CKUPBxo3GRIyixhVHrDgPfAknjAgwRkyy"
api = genius.Genius(client_access_token)

Loading in my list of rappers names from my previous notebook

with open('./artist_names.pkl', 'rb') as f:
    artist_names = pickle.load(f)

Chaning the name of artist ASAP Rocky to A$AP Rocky so genius can find it

artist_names[21] = 'A$AP Rocky'

Creating a function that for each artist's song will grab the album name, title of song, the artist, the lyrics, and the year it was released and retun it is a dictionary. 

def song_dict(song):
    out = {}
    for key in ['artist','album','title','lyrics','year']:
        out[key] = song.__getattribute__(key)
    return out

Running a for loop that for each artist in artist_names, will search the artist and create a list of dictionaries where each dictionary represents a song's data

for idx, name in enumerate(artist_names[36:]):
    try:
        artist_object = api.search_artist(name)
        artist_object = [
                song_dict(song) for song in artist_object.songs
            ]
        with open(f'./artist_objects/{name}.json', 'w') as f:
            json.dump(artist_object, f)
    except:
        print(idx, name, 'error')
        break
    time.sleep(.04)
