==> notebooks/03_Creating_dataframe.ipynb <==
# Creating the dataframe for analysis

In this notebook, I am going to iterate through all of the jsons for each artist and create a temporary dictionary where the columns are the keys and the values are the resulting list from each json. There will be columns for lyrics, year, title of song, and artist.

import json
import pandas as pd
import os

Creating an empty list of dataframes to be filled iteratively

list_of_dataframes=[]

Running a for loop that will create a dataframe for each song and append that dataframe to a list. The dataframes in this list will then be calculated into a f

for json_file in os.listdir('./artist_objects'):
    name = json_file.split('.')[0]
    with open(f'./artist_objects/{json_file}', 'r') as file:
        name_json = json.load(file)
    tmp = {
        'lyrics':[song['lyrics'] for song in name_json],
        'year':[song['year'] for song in name_json],
        'title':[song['title'] for song in name_json],
        'artist':[name for song in name_json]
    }
    list_of_dataframes.append(pd.DataFrame(tmp))
    
df = pd.concat(list_of_dataframes)
    
    

df.shape

df.info()

df.head(20)

df.to_csv('./Rappers.csv')
