==> notebooks/04_Cleaning_dataframe.ipynb <==
### Cleaning the dataframe

In this notebook, I am going to clean up my dateframe by mainly using regex to get rid of unwanted characters as well as removing duplicates

import pandas as pd
from difflib import SequenceMatcher as sm # For comparing similarity of lyrics
import regex as re
from nltk.tokenize import RegexpTokenizer

Reading in my newly constructed dataframe

df = pd.read_csv('./Rappers.csv', index_col=0, parse_dates=['year'])

df.info()

Dropping rows where there are nulls in the lyrics columnn. I lose $1539$ rows after doing this

len(df[df['lyrics'].isnull()].index)

df = df[df['lyrics'].isnull() == False]

df.reset_index(inplace=True)

Getting rid of words/characters from lyrics that are contained inside brackets, curly braces, or parentheses. I'm also getting rid of newline characters

pat = "\\[.+?\\]|\\(.+?\\)|\\{.+?\\}"
re_pat = re.compile(pat)
for i in range(len(df)):
    df.loc[i,'lyrics'] = re_pat.sub("", df.loc[i,'lyrics'])

pat = "\\n+"
re_pat = re.compile(pat)
for i in range(len(df)):
    df.loc[i,'lyrics'] = re_pat.sub(" ", df.loc[i,'lyrics'])

df.drop('index',axis=1,inplace=True)

df.reset_index(inplace=True)

df.drop('index',axis=1,inplace=True)

Dropping rows that have lyrics that are an empty string

for i in range(len(df)):
    if df.loc[i,'lyrics'] == '':
        df.drop(index=i,inplace=True)

df.reset_index(inplace=True)

df.drop('index',inplace=True,axis=1)

Creating a function that makes use of the sequence matcher library. This function compares to strings and returns a ratio indicating how similar they are. If the ratio is above $.3$, it returns a true. This function will be used to remove duplicates

def songsAreSame(s1, s2):    
    seqA = sm(None, s1, s2)
    seqB = sm(None, s2, s1)
    return seqA.ratio() > 0.3 or seqB.ratio() > 0.3

Creating a similar function but it compares titles to see if they are the same. The threshold in this function is $.5$

def titlesAreSame(s1, s2):    
    seqA = sm(None, s1, s2)
    seqB = sm(None, s2, s1)
    return seqA.ratio() > 0.5 or seqB.ratio() > 0.5

Dropping rows that are duplicates by comparing the titles and lyrics of rows that are next to each other (all duplicates appear next to each other). Dropping duplicates results in losing $6569$ rows. 

to_drop = []
for i in range(len(df)-1):
    if songsAreSame(df.loc[i,'lyrics'],df.loc[i+1,'lyrics']) or titlesAreSame(df.loc[i, 'title'],df.loc[i+1,'title']):
        to_drop.append(i+1)

len(to_drop)

for i in to_drop:
    df.drop(index=i,inplace=True)

Replacing the names of some of the artists with the correct punctuation manually 

df.replace('J Cole', 'J. Cole', inplace=True)

df.replace('TI', 'T.I.', inplace=True)

df.replace('The Notorious BIG', 'Notorious B.I.G.', inplace=True)

df.replace('DMC', 'Run DMC', inplace=True)

df.replace('NWA', 'N.W.A', inplace=True)

df.replace('BoB', 'B.o.B', inplace=True)

df.replace('RA The Rugged Man', 'R.A. The Rugged Man', inplace=True)

df.replace('Sir Mix-a-Lot', 'Sir Mix-A-Lot', inplace=True)

df.drop('index',inplace=True,axis=1)

df.reset_index(inplace=True)

Checking my artists to make sure their names are spelled correctly and with the correct punctuation / characters

df['artist'].unique()

df.to_csv('./clean_df')
