==> notebooks/05_Feature_engineering.ipynb <==
# Feature Engineering

In this notebook, I will create my target variable (whether or not the song appears on the billboard hot 100's) by seeing if the song exists in either of two billboard hot 100 dataframes found on the internet. 

import pandas as pd
import numpy as np
import math
from difflib import SequenceMatcher as sm # For comparing similarity of lyrics

Loading in my newly cleaned dataframe as well as my billboard data

df = pd.read_csv('./clean_df',index_col=0)

bb = pd.read_csv('./songs_1970_2018_uniq.csv', index_col=0, engine='python', parse_dates=['date'])

billboard = pd.read_csv('./Billboard_Hot100_Singles_1955-2013.csv', parse_dates=['Date'], skiprows=38)

billboard.set_index('Date',inplace=True)

Examining all three of my dataframes

df.head()

billboard.head()

bb.head()

Creating a column that has just the year as opposed to the full date

df.columns = ['lyrics', 'date', 'title', 'artist']

df['year'] = pd.to_datetime(df['date'], errors='coerce')

df['year'] = df.apply(lambda x: x['year'].year,1)

df.head()

Creating a function that will compare two strings and return true if they are similar but not equal (testing if it's mispelled)

def stringsAreMispelled(s1, s2):    
    seqA = sm(None, s1, s2)
    seqB = sm(None, s2, s1)
    return (seqA.ratio() > 0.7 or seqB.ratio() > 0.7) and s1 != s2

Creating a function that will return true if two strings are similar or equal

def stringsAreSame(s1, s2):    
    seqA = sm(None, s1, s2)
    seqB = sm(None, s2, s1)
    return (seqA.ratio() > 0.7 or seqB.ratio() > 0.7) or s1 == s2

Creating a funciton that returns true if the two strings are equal

def stringsAreEqual(s1, s2):    
    seqA = sm(None, s1, s2)
    seqB = sm(None, s2, s1)
    return s1 == s2

Renaming my artists and title columns in my two billboard dataframes so I can merge on them 

billboard.rename(columns={'Artist - Modified': 'artist', 'Title - Modified': 'title_bboard'}, inplace=True)

bb.rename(columns={'title': 'title_bboard'}, inplace=True)

Appending the two billboard dataframes together, one ends at 2013 so I'll subset the one that contains 2014 and and after to only include those years

bb_subset = bb[bb.index >= '2014']

final_bb = billboard.append(bb_subset, ignore_index=False)

Testing my final billboard dataframe with A$ap Rocky

final_bb[final_bb['artist'] == 'A$AP Rocky']

Creating a dataframe that from both my cleaned dataframe and my bilboard dataframe. By merging on artist and keeping the titles of the songs from both dataframes, I am able to get a dataframe that looks like this (see merged.head below)

df2 = df

merged = df2.merge(final_bb, on='artist',how='inner')[['artist', 'title', 'title_bboard']]

Instatiating my target variable column with all zeros

merged['hot100'] = 0

merged.head()

The column title has the title of every song A\$AP Rocky did, and it repeats that song's title for however many songs A\$AP Rocky had on the billboard hot100's. He had 4 songs, so the title is repeated 4 times. This happens for every song, so for each row I can compare each artist's song to all of the artist's hits to assess if the song was a hit or not. This column (hot100) will end up being my target variable, so doing this accurately is crucial to my model's success.

Creating masks for titles that are the same (similar or equal), songs that have similar titles (or are 'mispelled'), and songs that have only equal titles found in my dataframe and the billboard charts.

%time misspelled = merged.apply(lambda row: stringsAreMispelled(row['title'], row['title_bboard']), 1)

%time equal = merged.apply(lambda row: stringsAreEqual(row['title'], row['title_bboard']), 1)

%time same = merged.apply(lambda row: stringsAreSame(row['title'], row['title_bboard']), 1)

Filtering my merged dataframe with song titles that are found in both my dataframe and on the billboard charts.

hot100 = merged[same]

$893$ hits were found

len(hot100)

hot100.reset_index(inplace=True)

Creating a column that represents if the song was a hit or not. I instantiate it will all zeroes then I iterate through the rows and see if the title of the song was found on the billboard hot 100's.  

merged['hot100'] = 0

for i in merged[same].index:
    merged.loc[i,'hot100'] = 1

The number of hits found lines up with my hot100 dataframe

sum(merged['hot100'])

Checking my dataframe with its new column, hot100

merged.head()

merged[same].head()

Now I will create my final dataframe, by merging only the hits onto the dataframe. The songs that are not hits will have NaN values in this column (which I'll label as hot100). I'll fill these nan's with zeros and they'll be my positive class

df_final = df.merge(merged.iloc[merged[one].index, :][['artist', 'title', 'hot100']].drop_duplicates(), how='left')

df_final.fillna(0, inplace=True)

sum(df_final['hot100'])

My positive class is 870 long. This unforutantely means that my classes are severely unbalanced, as my negative class has more than 15,000 observations. This will have to be dealt with later, but for now I must manually comb through my positive class where the titles are similar (and not equal) and see if my comparison function mislabeled a song as a positive class becuase its title was similar to that of a hit. If a song labeled as a hit is in fact not a hit, I'll have to relable it as zero. Instead of showing all of this, I will only show myself doing this one time so my process can be seen.

First I have to view a section of my dataframe to manually check for mispellings. I can see on the fourth row that the song 'Pull Up' was mistaken for the hit'Roll Up'. Since these are in fact not the same song, 'Pull Up' will have to be relabeled as a zero.

merged[misspelled][150:300]

This song was an advertisement but got mistaken for a hit  so it must be dropped. 

df_final.drop(index = 10264, inplace=True) # advertisement

Giving the song 'Still Drake' a zero since it was not a hit.

df_final.loc[22081,'hot100'] = 0

After I checked all mispelled songs (about 300 in total), I ended up removing only about $38$ songs from my positive class

df_final['hot100'].sum()

Saving this dataframe as df_hot100

df_final.to_csv('./df_hot100.csv')
