==> notebooks/06_Filling_Missing_Years.ipynb <==
# Filling Missing Years for Hits

This notebook is an appendix to my data cleaning notebook. At this point in the data cleaning and engineering process, I am going to deal with all my missing dates data. Unfortunately, some of my positive class had observations had missing dates. Since my positive class is so small already, I must manually search for them online as I cannot afford to lose anymore obseravtions from this class. 

import pandas as pd
import google as google
import urllib.request
import webbrowser

Reading in my new dataframe that includes the classes

df = pd.read_csv('./df_hot100.csv', index_col = 0)

Some hits were missing lyrics so I manually grabbed them from online.

df.loc[21915,'lyrics'] = """Hermès link, ice blue mink Tat on my ribs like I do not know what permanent is They want me gone, wait for the kicker Bury me now and I only get bigger, that's word to my nigga Yeah, October Firm in the cut Stay at the top like I'm stuck, that's just how I'm giving it up She wanna get married tonight, but I can't take a knee 'cause I'm wearing all white Me and my broski are twins but we don't look alike I don't take naps Me and the money are way too attached to go and do that Muscle relax That and a spliff put me right on my back, I gotta unpack Virginia Black I could go make enough money off that and not even rap What's that? Facts Contract max, I gotta bring that shit back Hermes link, ice blue mink Tat on my ribs like I do not know what permanent is They want me gone, out of the picture Bury me now and I only get bigger, that's word to my, word to my I'm so hot yeah, I'm so right now Who's not gang bitch? Let me find out Keep hearing clicks when I'm talkin' on the iPhone Feds in the city hate to see us on a high note I switch flow like I switch time zone Can't get Nobu but you can get Milestone I got a do mansion, 'cause I outgrew condo Gotta do Maybach, she wan' fuck on the drive home Yeah, met her once and I got through I'm never washed but I'm not new I know I said top five, but I'm top two And I'm not two and I got on Thought you had one, but it's not one, nigga nah Hermès link, ice blue mink Tat on my ribs like I do not know what permanent is They want me gone, wait for the kicker Bury me now and I only get bigger, that's word to my nigga"""

df.loc[21915,'lyrics'][0:110]

len(df[df['year'] != 0])

Let's see how big my dataframe is when I filter it so only the observations with no date are present

df[(df['hot100'] == 1) & (df['year'] == 0)].shape

Now I am going to manually look up the dates for all 50 of these observations and find their dates and fill them in manually. This is tedious, but crucial to my model's performance.

df.loc[1914,'year'] = 2001

df.loc[1912,'year'] = 2004

df.loc[1409,'year'] = 2011

df.loc[1369,'year'] = 2002

df.loc[1352,'year'] = 2010

df.loc[1333,'year'] = 2000

df.loc[1296,'year'] = 2016

df.loc[594,'year'] = 2012

df.loc[25098,'year'] = 2012

df.loc[24041,'year'] = 1993

df.loc[23988,'year'] = 1995

df.loc[22434,'year'] = 1998

df.loc[22394,'year'] = 1997

df.loc[22081,'year'] = 2009

df.loc[21915, 'year'] = 2017

df.loc[21713,'year'] = 2004

df.loc[21485,'year'] = 2012

df.loc[20922,'year'] = 1996

df.loc[20604,'year'] = 2008

df.loc[18166,'year'] = 1999

df.loc[20534,'year'] = 2004

df.loc[19639,'year'] = 1996

df.loc[19141,'year'] = 1993

df.loc[18788,'year'] = 2012

df.loc[18106,'year'] = 2012

df.loc[17534, 'year'] = 2003

df.loc[17013,'year'] = 1999

df.loc[16921,'year'] =2004

df.loc[11131,'year'] = 2007

df.loc[2773,'year'] = 1996

df.loc[10163,'year'] = 2001

df.loc[10107,'year'] = 200

df.loc[9995,'year'] = 1999

df.loc[9184,'year'] = 1999

df.loc[8183,'year'] = 2013

df.loc[4431,'year'] = 1998

df.loc[3225,'year'] = 1996

df.loc[16860,'year'] = 1996

df.loc[16860,'year'] = 2010

df.loc[16812,'year'] = 2005

df.loc[16697,'year'] = 1995

df.loc[14579,'year'] = 2002

df.loc[2222,'year'] = 2013

df.loc[2438,'year'] = 2013

df.loc[11967,'year'] = 2003 

df.loc[12209,'year'] = 2010

df.loc[13235,'year'] = 2010

df.loc[13272,'year'] = 2003

df.loc[13395,'year'] = 2008

df.loc[13416,'year'] = 2004

df.loc[10107,'year'] = 1999

df.loc[6720,'year'] = 2000

df = df[df['year'] !=0]

df.reset_index(inplace=True)

df.drop('index',axis=1,inplace=True)

Saving my new cleaned data frame. This is my final step in the cleaning/engineering process and now I can go on to tokenizing my data and modeling

df.to_csv('./clean_data_years.csv')
