==> notebooks/07_tokenizing_local.ipynb <==
# Tokenizing 

In this notebook I am first going to lemmatize my data, and then tokenize it by using two methods. One will be term frequency inverse docuemtn frequency transformer, and the other method will be cout vectorizing. I will use the count vectorized dataframe mainly for eda, as it simply counts the appearance of a word in a document (ourd document's will be rap songs) and thus is easily interpreted. I will reserve my TFIDF data for modeling. 

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer, TfidfTransformer, CountVectorizer
import numpy as np
from sklearn.feature_extraction import stop_words
import nltk
from nltk.stem import WordNetLemmatizer
import re
import time

Reading in my recently cleaned data

df = pd.read_csv('./clean_data_years.csv', index_col=0)

df.head(7)

Instantiating my Word Net Lemmatizer, which I'll use for lemmatizing 

wnt = WordNetLemmatizer()

Creating a function that will lowercase my lyrics,get rid of quoutes and backslashs, as well remove 'ing' and its variants. This will ensure the lemmatization goes smoothly and works properly.

def lemmatize(df):
        df['lyrics'] = df['lyrics'].str.lower()
        df['lyrics'] = df['lyrics'].str.replace('"', '')
        df['lyrics'] = df['lyrics'].str.replace('(,|\.)', '')
        df['lyrics'] = df['lyrics'].map(lambda x: re.sub("(in|ing|in')", '', x))
        
        df['lyrics'] = df['lyrics'].map(lambda x: ' '.join([wnt.lemmatize(word) for word in x.split()]))

Lemmatizing the data

lemmatize(df)

Checking my work

df.head()

Saving my lemmatized data

df.to_csv('./final_clean_lemma_df.csv')

Creating a custom list of stop words to not include in my tokenized data frames. A lot of things I'm excluding are things parts of words with punctuation, like the 'll' in 'I'll'. This is done because tokenizers tokenize on punctuation.

custom_stopwords = list(stop_words.ENGLISH_STOP_WORDS)

custom_stopwords.extend(['like', 'll', 'ain','don','em','er','wa',
                         'ya','just','let','got','den','ol','izz','im',
                         'letting','hol','right','hah','dat','ve','mon',
                         'la', 'aw','whit','ma','da','uhh','gon','wit'])

Tokenizing my dataframe. I'm excluding words that don't appear in at least $5$ documents and excluding words that appear in over $70\%$ of the documents. This will hopefully tame the size of my tokenized dataframes

cvec = CountVectorizer(stop_words=custom_stopwords, min_df=5,max_df=.7)

tvec = TfidfVectorizer(stop_words=custom_stopwords, min_df=5,max_df=.7)

Xc = cvec.fit_transform(df['lyrics'])

Xt = tvec.fit_transform(df['lyrics'])

Creating tokenized dataframes out of my tokenized arrays to be used for modeling and EDA

df_token = pd.DataFrame(Xt.toarray(), columns=cvec.get_feature_names())

df_token['date_year'] = df['year']

X = pd.DataFrame(Xc.toarray(),columns=cvec.get_feature_names())

X['date_year'] = df['year']

Saving my dataframes

X.to_csv('./X.csv')

df_token.to_csv('./df_token.csv')

df_token.shape
